+++ example/example-script.as
@@ -1,0 +1,13 @@
+游戏脚本示例&
+/*这是一段注释*/
+{
+    这是文本A
+    #选项一{
+        这是文本B
+        #退出
+    }
+    #选项三{
+        这是文本C
+        #退出
+    }
+}
