--- script.as
@@ -1,9 +1,0 @@
-游戏脚本示例&
-/*这是一段注释*/
-{这是文本A
-#选项一
-    {这是文本B
-    #选项二
-    }
-#选项三
-}
